+++ lib/compendium/count_query.rb
@@ -1,9 +1,16 @@
+require 'compendium/errors'
 require 'compendium/query'
 
 module Compendium
   # A CountQuery is a Query which runs an SQL count statement
   # Often useful in conjunction with a grouped query
   class CountQuery < Query
+
+    def initialize(*args)
+      super
+
+      @options.reverse_merge!(order: 'COUNT(*)', reverse: true)
+    end
 
   private