+++ spec/sum_query_spec.rb
@@ -9,13 +9,36 @@ class SingleSummer
 end
 
 class MultipleSummer
+  def order(*)
+    @order = true
+    self
+  end
+
+  def reverse_order
+    @reverse = true
+    self
+  end
+
   def sum(col)
-    { 1 => 123, 2 => 456, 3 => 789 }
+    results = { 1 => 340, 2 => 204, 3 => 983 }
+
+    if @order
+      results = results.sort_by{ |r| r[1] }
+      results.reverse! if @reverse
+      results = Hash[results]
+    end
+
+    results
   end
 end
 
 describe Compendium::SumQuery do
   subject { described_class.new(:counted_query, :col, { sum: :col }, -> * { @counter }) }
+
+  it 'should have a default order' do
+    subject.options[:order].should == 'SUM(col)'
+    subject.options[:reverse].should == true
+  end
 
   describe "#run" do
     it "should call sum on the proc result" do
@@ -29,9 +52,21 @@ describe Compendium::SumQuery do
       subject.run(nil, self).should == [1792]
     end
 
-    it "should return a hash if given" do
-      @counter = MultipleSummer.new
-      subject.run(nil, self).should == { 1 => 123, 2 => 456, 3 => 789 }
+    context 'when given a hash' do
+      before { @counter = MultipleSummer.new }
+
+      it "should return a hash if given" do
+        subject.run(nil, self).should == { 3 => 983, 1 => 340, 2 => 204 }
+      end
+
+      it 'should be ordered in descending order' do
+        subject.run(nil, self).keys.should == [3, 1, 2]
+      end
+
+      it 'should use the given options' do
+        subject.options[:reverse] = false
+        subject.run(nil, self).keys.should == [2, 1, 3]
+      end
     end
 
     it "should raise an error if the proc does not respond to sum" do